String Manipulation to SimpleXML and DOMDocument
I removed the string manipulation in favor of something more development friendly.  The drawback is that the resulting XML is not as pretty, although I did run the processed fields through DOMDocument pretty printer.  The entire resultant XML could be run through a pretty printer, but still won't achieve the tab indented node attributes.

The entire process could be made much more efficient, but this was just an initial conversion.